requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow
requests
Pillow
xgboost
